fakeroot: disallow selection for the target
fakeroot for the target fails to build against glibc, due to largefile
related problems :

libtool: link: /home/test/outputs/test-503/host/usr/bin/powerpc-linux-gnu-gcc -pipe -Os -D_LARGEFILE_SOURCE -D_LARGEFILE64_SOURCE -D_FILE_OFFSET_BITS=64 -o faked faked.o  ./.libs/libcommunicate.a -ldl
{standard input}: Assembler messages:
{standard input}:2592: Error: symbol `__fxstatat64' is already defined
{standard input}:2627: Error: symbol `__fxstat64' is already defined
{standard input}:2663: Error: symbol `__xstat64' is already defined
{standard input}:2698: Error: symbol `__lxstat64' is already defined
make[3]: *** [libfakeroot.lo] Error 1

As fakeroot for the target is not really needed, and was only added
when we packaged fakeroot for the host, just make fakeroot not
selectable for the target.

We have to keep the AUTOTARGETS call for fakeroot on the target,
because the host variant does not work otherwise, due to issues in the
package infrastructure (that will be fixed post-2011.05).

Signed-off-by: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>
Signed-off-by: Peter Korsgaard <jacmet@sunsite.dk>