pppd: fix plugin search path
Fix the plugin search path by passing a --prefix to configure.
DESTDIR is used in the source code for the search path of plugins
(DESTDIR/lib/pppd/VERSION).
Since DESTDIR defaults to /usr/local we get the search path wrong so the
'plugin' directive requires a full pathspec to the plugin rather than
the .so name, which isn't nice or portable.

Signed-off-by: Gustavo Zacarias <gustavo@zacarias.com.ar>
Signed-off-by: Peter Korsgaard <jacmet@sunsite.dk>