docs: rework LIBFOO_SITE_METHOD documentation
The current structure of the LIBFOO_SITE_METHOD paragraph would make
it difficult to add documentation for the "file" and "local" methods.
There are enough possible values for LIBFOO_SITE_METHOD that
describing them in a sub-list seems to work well, which makes it
easier to add documentation for new methods.

Signed-off-by: Nathan Lynch <ntl@pobox.com>
Acked-by: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>
Acked-by: Arnout Vandecappelle (Essensium/Mind) <arnout@mind.be>
Signed-off-by: Peter Korsgaard <jacmet@sunsite.dk>