xserver_xorg-server: fix dbus/udev configure problem
If both the dbus and udev packages are selected, then the configure step
for the xserver_xorg-server package fails as follows.

  configure: error: Hotplugging through both libudev and dbus/hal not allowed

The configure script no longer allows both the --enable-config-dbus
and --enable-config-udev options to be specified.

This patch changes the xserver_xorg-server makefile to use udev in preference
to dbus, when both are available.

Signed-off-by: Simon Dawson <spdawson@gmail.com>
Signed-off-by: Peter Korsgaard <jacmet@sunsite.dk>