perl: remove the double configure hack
configure had to be called twice because with --mode=cross the
miniperl-step failed.  However, just leaving out the --mode parameter
is sufficient to make it work.  Since GNU_TARGET_NAME is always
different from the host's tuple (it has -buildroot- in it), we can
safely assume that the configure script will automatically enter
cross mode.

Also fix a type in perladmin definition.

Signed-off-by: Arnout Vandecappelle (Essensium/Mind) <arnout@mind.be>
Signed-off-by: Peter Korsgaard <jacmet@sunsite.dk>