matchbox-desktop: fix link issue against libdl
matchbox-desktop uses dlopen(), so it needs to link against libdl. The
configure.ac and Makefile.am have the necessary provisions to do so,
but the included configure script does not replace LIBADD_DL in
src/Makefile.am as it should.

Therefore, we force the autoreconf of the package, which solves the
problem.

Signed-off-by: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>
Signed-off-by: Peter Korsgaard <jacmet@sunsite.dk>