dvbsnoop: new package dvbsnoop
dvbsnoop is a DVB/MPEG stream analyzer. It's designed to debug, dump or
view the digital stream info available via

    satellite   -> DVB-S, DVB-S2
    cable       -> DVB-C, DVB-C2
    terrestrial -> DVB-T, DVB-T2

dvbsnoop can be used on any digital settopbox that running on linux and
provides the DVB APIv3. It's also usable on classical PCs that contains
a DVB hardware (PCI Card or USB Plug).

[Peter: needs largefile, fix help text]
Signed-off-by: Carsten Schoenert <c.schoenert@gmail.com>
Signed-off-by: Peter Korsgaard <jacmet@sunsite.dk>