foomatic-filters: make location of enscript very clear
Passing --with-file-converter=enscript isn't enough since the build
system will try the fallback options (a2ps, mpage, texttops) if it's not
found.
This has two bad effects: one of the tools leaking from the distribution
making foomatic-filters think everything will be the same on the target
(and failing when that's not the case), and failed builds when none is
found (enscript is installed in the target directory, the build system
doesn't look there).
So just avoid the test by stating explicitly where enscript will live in
the target. Fixes:
http://autobuild.buildroot.net/results/37ad3d2c5c35f00c644a2c572f1429bafa4b7cf4/

Signed-off-by: Gustavo Zacarias <gustavo@zacarias.com.ar>
Signed-off-by: Peter Korsgaard <jacmet@sunsite.dk>