pulseaudio: link against json-c
Add an upstream patch that makes Pulseaudio link against json-c
instead of json, so that it works with json-c 0.11, and doesn't cause
problems with libjson has been also installed in the system. Note that
this fix has been merged in Pulseaudio after the 3.0 release, so we
will have to keep this patch around when bumping to 3.0.

Fixes
http://autobuild.buildroot.org/results/91ffd3196092c48b88f59adb12741b3f93064dea/build-end.log

Signed-off-by: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>
Signed-off-by: Peter Korsgaard <jacmet@sunsite.dk>