configs: bump kernel use for RPi, use simpler defconfig
Bump for a slew of USB-related and other fixes.

The bcmrpi defconfig is a huge beast with lots of superfluous drivers.
The brcmrpi_quick defconfig is a stripped down minimalist defconfig with
all that is needed to bring up a RPi with complete HW support, and
nothing superfluous, which is what we need for a board's defconfig in
buildroot.

Signed-off-by: "Yann E. MORIN" <yann.morin.1998@free.fr>
Signed-off-by: Peter Korsgaard <jacmet@sunsite.dk>