tcping: new package
tcping does a TCP connect to the given ip/port combination. The user can
specify a timeout in seconds. This is useful in shell scripts running in
firewalled environments. Often SYNs are just being dropped by firewalls,
thus connection establishment will be retried several times (for minutes)
until a TCP timeout is reached. With tcping it is possible to check first
if the desired port is reachable and then start connection establishment.

http://www.linuxco.de/tcping/tcping.html

[Peter: wrap help text, pass TARGET_LDFLAGS]
Signed-off-by: Arnaud Rébillout <rebillout@syscom.ch>
Signed-off-by: Peter Korsgaard <jacmet@sunsite.dk>