libnl: simplify makefile by using --{enable, disable}-cli
The libnl configure script has a --{enable,disable}-cli options that
allows to enable or disable the compilation of the libnl tools. Use
this option instead of compiling everything and then removing the
installed programs.

Note that we also get rid of the uninstall command, which is planned
to be globally removed in Buildroot.

Signed-off-by: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>
Acked-by: "Yann E. MORIN" <yann.morin.1998@free.fr>
Reviewed-by: Samuel Martin <s.martin49@gmail.com>