qt5: Disable on MIPS64 BigEndian n64 platforms
qt5base fails to compile on MIPS64 BigEndian n64 platforms because there
are multiple calls to some functions that are disabled. Since every Qt5
package depends on qt5base we can disable Qt5 entirely. This is a quick
fix for the next release. In the meantime, a proposed patch to fix this
problem has already been sent to Qt:

https://bugreports.qt-project.org/browse/QTBUG-35228

Signed-off-by: Vicente Olivert Riera <Vincent.Riera@imgtec.com>
Acked-by: Markos Chandras <markos.chandras@imgtec.com>
Signed-off-by: Peter Korsgaard <peter@korsgaard.com>