supervisor: convert to the Python package infrastructure
Remove the select of BR2_PACKAGE_PYTHON_SETUPTOOLS, since what is
really needed is host-setuptools (as can be seen from the original .mk
file), and not setuptools for the target.

Also, remove the mangling of the Python shebang, since it's now done
by passing the --executable= option at install time (done in the
Python package infrastructure).

Finally, convert the package to use <pkg>_INSTALL_INIT_SYSV.

Signed-off-by: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>
Signed-off-by: Peter Korsgaard <peter@korsgaard.com>