rtai: install to staging
Building RTAI userspace applications requires having the RTAI headers
and libraries installed to the staging directory, so this patch adds
RTAI_INSTALL_STAGING = YES.

In addition, this patch fixes up the rtai-config script, but does that
manually, because the rtai-config script doesn't match the convention
that the common <pkg>_CONFIG_SCRIPTS logic is designed for. It also
specifies a special directory for the RTAI headers, so that they don't
get mixed up with kernel headers.

Finally, it updates the RTAI download address to use the new correct
location.

[Thomas: fix commit log.]

Signed-off-by: Andrew Barnes <andy@outsideglobe.com>
Signed-off-by: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>