configs: add SoCkit defconfig
This commit introduces initial support for the Arrow SoCkit
boards, featuring an Altera Cyclone V SoC:
http://www.altera.com/b/arrow-sockit.html

It is based on Roman Diouskine's work at
https://github.com/rndi/buildroot-alt.

The kernel and the u-boot used is the one developped by the
rocketboards.org community, as mainline support is not fully
available yet.

More info is available in the board/altera/sockit/readme.txt file.

[Thomas: remove a number of not really needed empty new lines in the
readme.txt file, update the defconfig for the new kernel headers
options.]

Signed-off-by: Sebastien Bourdelin <sebastien.bourdelin@savoirfairelinux.com>
Signed-off-by: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>