xbmc: Fix Java dependency check
A check for Java has been implemented in dependencies.sh, but during the
review, the name of the variable had been changed from:
BR2_HOST_NEEDS_JAVA to BR2_NEEDS_HOST_JAVA
as it is more accurate.

The change was not made in XBMC and as such, you could start the build
of XBMC without Java on your system.

Signed-off-by: Maxime Hadjinlian <maxime.hadjinlian@gmail.com>
Acked-by: Bernd Kuhls <bernd.kuhls@t-online.de>
Signed-off-by: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>