package/mesa3d: fix improper dependency of blind option
Currently, the blind option BR2_PACKAGE_MESA3D_DRI_DRIVER depends
on !STATIC.

But this option is also selected by the various DRI drivers, and
none of them currently depend on !STATIC (although there is a comment
stating DRI drivers need !STATIC, there's nothing to enforce that).

So, we could well end-up with an inconsistent configuration, where some
DRI drivers are selected even though STATIC is set.

Enclose all DRI drivers in an 'if !STATIC' condition, remove the
dependency from the blind option, move the comment so it is nearer
the affected drivers, rephrase the comment to match the rules about
dependencies on toolchain features.

[Thomas: really use the right wording for the comment about the
dynamic library dependency.]

Signed-off-by: "Yann E. MORIN" <yann.morin.1998@free.fr>
Cc: Paul Cercueil <paul@crapouillou.net>
Cc: Bernd Kuhls <berndkuhls@hotmail.com>
Signed-off-by: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>