sane-backends: uses fork(), not available on noMMU systems
sane-backends currently doesn't build with the ADI external
toolchains, due to an internal compiler failure. However, using the
internal toolchain backend for Blackfin goes past this internal
compiler failure, but raises the problem that sane-backends uses
fork(). Sane-backends has a --enable-pthreads option that allows to
use threads instead of forking in some parts of sane, but there are
still some other areas of sane-backends that continue to use fork().

Since the probability of using sane-backends on a noMMU system is
fairly small, we'll just disable this package for now on such
architectures, until someone cares enough to investigate deeper.

Fixes:

  http://autobuild.buildroot.org/results/136/136100c3a2d843b7915fdd83d23d8b301fee577e/

Signed-off-by: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>
Signed-off-by: Peter Korsgaard <peter@korsgaard.com>