isl: new package
isl is a library for manipulating sets and relations of integer points
bounded by linear constraints.

isl is needed for the optional graphite optimisations that are supported
by gcc since version 4.5. Therefore this package is required for the
toolchain to support graphite.

Graphite optimisations primarily involve loop blocking flattening and
interchage so are probably of mimimal use in an embedded system where
small sizes are favoured.

[Thomas:
 - bump to 0.12.2 (cannot use 0.13, incompatible with cloog 0.18.2,
   comment added about this)
 - use .bz2 tarball
 - disable libtool patch, which doesn't apply]

Signed-off-by: Steve Thomas <scjthm@live.com>
Signed-off-by: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>