mpg123: correct powerpc cpu variant selection
Fixes http://autobuild.buildroot.net/results/d8a/d8af8f991b4429cf06081ddf2f08043861842a50/

ppc_nofpu should (as the name indicates) only be used for powerpc.

It is not clear to me if this code works for powerpc64(le) as well, so this
is only selected for classic ppc32.

Signed-off-by: Peter Korsgaard <peter@korsgaard.com>