faifa: force GIT_REV to a correct value
This variable is looked up by the Makefile throught a git command line
invocation.
Since we don't keep the '.git' folder in the packages when we clone
them, the command line fail to find anything.

Even worse, it can keep going up the tree until it finds buildroot's
'.git' directory.

Signed-off-by: Maxime Hadjinlian <maxime.hadjinlian@gmail.com>
Signed-off-by: Peter Korsgaard <peter@korsgaard.com>