package/sispmctl: fix static linking
The static linking doesn't work with libusb-compat when using
libusb-config. Also the CFLAGS initialization with libusb-config
--cflags is comented out.

Based on libftdi's patch, this commit adds the support for pkg-config
to link sispmctl correctly.

This allow to add the correct include directories to CFLAGS (Add
-I$(STAGING_DIR)/usr/include and
-I$(STAGING_DIR)/usr/include/libusb-1.0) and also add the required
libraries during static linking (-lusb -lusb-1.0 and -pthread).

Since configure.ac is modified, the configure script needs to be
regenerated with autoreconf.

Fixes:
http://autobuild.buildroot.net/results/647/6473088e751d3ab3a5227e9d7876966934e66378/build-end.log

Cc: Phil Eichinger <phil.eichinger@gmail.com>
Signed-off-by: Romain Naour <romain.naour@openwide.fr>
Signed-off-by: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>
Tested-by: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>