package/freerdp: bump version
There has been no new release recently, and the 1.0 branch is out-dated.

The 1.1 branch has not been release-tagged since the last beta tag,
but it still continues to receive bug fixes, and is relatively stable.

The next major release should be 1.2.0, but it is still in beta, looks
like it is focused on Android, and was only recently tagged.

So, we use the latest cset from the 1.1 branch until there is a new
release (either 1.2.0 or 1.1.0), at which point we can revisit which
version we'll use.

Drop our patch, since the problem has been fixed upstream (with a more
complete solution.)

Signed-off-by: "Yann E. MORIN" <yann.morin.1998@free.fr>
Signed-off-by: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>