coreutils: fix libintl, reorder ENV, OPTS & PROGS
Move OPTS, ENV & PROGS up to the top, the fact that a conditional
(BR2_PACKAGE_BUSYBOX) was before them is bad style and confused me while
i tried to get them grouped together.
This was making all of the new disable OPTS fail, pretty much harmless
but still.

Account for libintl/gettext presence, it's not required in any toolchain
combination but it will fail if it's present, so use it when it's
selected by some other package. Fixes:
http://autobuild.buildroot.net/results/30d/30d0e3d27d8606443479466e60716e6f202a4711/

Add conditional on OpenSSL for faster hashing binaries as pointed again
by Pádraig Brady.

Signed-off-by: Gustavo Zacarias <gustavo@zacarias.com.ar>
Signed-off-by: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>