mp4v2: disable CFLAGS mangling
Fixes
http://autobuild.buildroot.net/results/235/235cade242ce3bcd583c57575a3612b811fa6e89/
http://autobuild.buildroot.net/results/529/5298118b20f5e0c4c38104411d979e1c17cca04b/
http://autobuild.buildroot.net/results/773/773c53c9c7d422b319dc0d9196426cae06d81c14/
http://autobuild.buildroot.net/results/41e/41eb1936c3afe329d1321f5a0f071db098b2fb37/

And many others.

When mp4v2 is configured with --disable-debug (which the infrastructure will
pass unless BR2_ENABLE_DEBUG is enabled) then the configure script tries to
filter out any -g* debugging flags with the following logic:

if test "$enable_debug" = "no"; then
    changequote(<<,>>)dnl
    if test "$GCC" = "yes"; then
        CFLAGS=`echo "$CFLAGS" | sed -e 's/-g[^ ]*[ ]*//' -e 's/^ //' -e 's/ $//'`
    fi
    if test "$GXX" = "yes"; then
        CXXFLAGS=`echo "$CXXFLAGS" | sed -e 's/-g[^ ]*[ ]*//' -e 's/^ //' -e 's/ $//'`
    fi
    changequote([,])dnl
fi

This logic is unfortunately buggy as it also strips up arguments like
-mfloat-gprs=double breaking the build when the powerpc SPE variants are
-used.

Instead of fixing the sed invocation we can simply pass --enable-debug to
configure as the debugging flag is not used for anything else.

Signed-off-by: Peter Korsgaard <peter@korsgaard.com>